requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
ebooklib>=0.18
tqdm>=4.60.0
tiktoken>=0.4.0

# PDF OCR dependencies
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=9.0.0
pdf2image>=1.16.0
